+++ command/app.go
@@ -3,6 +3,7 @@ package command
 import (
 	"context"
 	"fmt"
+	"os"
 
 	cmpinstall "github.com/posener/complete/cmd/install"
 	"github.com/urfave/cli/v2"
@@ -113,6 +114,15 @@ var app = &cli.App{
 		// After callback is not called if app exists with cli.Exit.
 		parallel.Close()
 		log.Close()
+	},
+	OnUsageError: func(c *cli.Context, err error, isSubcommand bool) error {
+		if err != nil {
+			_, _ = fmt.Fprintf(os.Stderr, "%s %s\n", "Incorrect Usage:", err.Error())
+			_, _ = fmt.Fprintf(os.Stderr, "See 's5cmd --help' for usage\n")
+			return err
+		}
+
+		return nil
 	},
 	Action: func(c *cli.Context) error {
 		if c.Bool("install-completion") {